requirements.txt
datatest==0.11.1
Flask==1.1.2
Flask-Caching==1.9.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
